imsomnia refactor
