--- src/Front/index.js
@@ -1,9 +1,0 @@
-import { checkPlayList } from "./controller/playList/playList.js";
-
-var playList = [];
-var currentPlaylist = [];
-var shufflePlaylist = [];
-
-var repeat = false;
-var shuffle = false;
-
